BACKEND/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
prisma==0.15.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pydantic[email]==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.0

groq==0.14.0
paddleocr==2.8.1
paddlepaddle>=2.6.1
opencv-python-headless>=4.9.0.80
pdf2image==1.17.0
Pillow==11.0.0
